+++ src/rime/dict/reverse_lookup_dictionary.h
@@ -21,7 +21,7 @@ namespace rime {
 namespace reverse {
 
 struct Metadata {
-  static const int kFormatMaxLength = 32;
+  static const size_t kFormatMaxLength = 32;
   char format[kFormatMaxLength];
   uint32_t dict_file_checksum;
   String dict_settings;
