+++ src/rime/gear/speller.h
@@ -41,7 +41,7 @@ class Speller : public Processor {
   string delimiters_;
   string initials_;
   string finals_;
-  int max_code_length_ = 0;
+  size_t max_code_length_ = 0;
   bool auto_select_ = false;
   bool use_space_ = false;
   boost::regex auto_select_pattern_;
