+++ src/rime/gear/table_translator.h
@@ -43,8 +43,8 @@ class TableTranslator : public Translator,
   bool enable_sentence_ = true;
   bool sentence_over_completion_ = false;
   bool encode_commit_history_ = true;
-  int max_phrase_length_ = 5;
-  int max_homographs_ = 1;
+  size_t max_phrase_length_ = 5;
+  size_t max_homographs_ = 1;
   the<Poet> poet_;
   the<UnityTableEncoder> encoder_;
 };
